+++ spec/json/matchers/match_response_schema_spec.rb
@@ -14,23 +14,57 @@ describe JSON::Matchers, "#match_response_schema" do
   end
 
   it "fails when the body is missing a required property" do
-    create_schema("array_schema", {
+    create_schema("foo_schema", {
       type: "object",
       required: ["foo"],
     })
 
-    expect(response_for({})).not_to match_response_schema("array_schema")
+    expect(response_for({})).not_to match_response_schema("foo_schema")
   end
 
   it "fails when the body contains a property with the wrong type" do
-    create_schema("array_schema", {
+    create_schema("foo_schema", {
       type: "object",
       properties: {
         foo: { type: "string" }
       }
     })
 
-    expect(response_for({foo: 1})).not_to match_response_schema("array_schema")
+    expect(response_for({foo: 1})).not_to match_response_schema("foo_schema")
+  end
+
+  it "contains the body in the failure message" do
+    create_schema("foo", { type: "array" })
+
+    expect {
+      expect(response_for(bar: 5)).to match_response_schema("foo")
+    }.to raise_error(/{"bar":5}/)
+  end
+
+  it "contains the body in the failure message when negated" do
+    create_schema("foo", { type: "array" })
+
+    expect {
+      expect(response_for([])).not_to match_response_schema("foo")
+    }.to raise_error(/\[\]/)
+  end
+
+  it "contains the schema in the failure message" do
+    schema = { type: "array" }
+    create_schema("foo", schema)
+
+    expect {
+      expect(response_for(bar: 5)).to match_response_schema("foo")
+    }.to raise_error(/#{schema.to_json}/)
+  end
+
+  it "contains the schema in the failure message when negated" do
+    schema = { type: "array" }
+    create_schema("foo", schema)
+
+    expect {
+      expect(response_for([])).not_to match_response_schema("foo")
+    }.to raise_error(/#{schema.to_json}/)
   end
 
   it "does not fail when the schema matches" do